pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "auditgraph"
version = "0.1.0"
description = "Local-first deterministic personal knowledge graph tooling"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Auditgraph Contributors" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
	"pyyaml>=6.0.3",
	"neo4j>=5,<6",
	"pypdf>=6.9.1",
	"python-docx>=1.2.0",
	"spacy>=3.7,<4",
	"dulwich>=0.22",
	"jsonschema>=4,<5",
	"markdown-it-py[linkify]>=4,<5",
]

[project.scripts]
auditgraph = "auditgraph.cli:main"

[project.urls]
Repository = "https://github.com/socratesone/auditgraph"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["auditgraph*"]

[tool.setuptools.package-data]
# Ship rule-pack stubs as package resources (Spec-028 #8 / adjustments2.md §8).
# The validator falls back to these via importlib.resources when a workspace-
# local copy is absent; scaffold.initialize_workspace copies them into new
# workspaces. Top-level config/ is an editable-checkout mirror, not the
# wheel-packaging mechanism.
auditgraph = ["_package_data/**/*.yaml", "_package_data/**/*.yml"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers"
markers = [
    "slow: end-to-end story tests; opt out via 'pytest -m \"not slow\"'",
]
